adhoc_jobs/git/myAgent-Shopping/AgentHub/pyproject.toml
[project]
name = "agora"
version = "0.1.0"
description = "Agent registration, capability declaration and discovery platform"
requires-python = ">=3.11"
dependencies = [
    "aiomqtt>=1.0.0",
    "paho-mqtt>=1.6.0",
    "pyyaml>=6.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.setuptools.packages.find]
include = ["agora*", "hub*", "agents*"]

[project.scripts]
agora-hub = "hub.main:main"
echo-agent = "agents.echo_agent.agent:main"
calc-agent = "agents.calculator_agent.agent:main"
